Fixes Cerestation powerissues
The engineering smes were sent 200k in and 190k out, with departmental smes at 95k in and 90k out, causing the four engineering smes to drain to empty in less then 10 minutes.
This pr resets those numbers to default 50 in 50 out, which will give overall better power, as each department starts with a 60% charged smes.